+++ src/setPreviewData.ts
@@ -1,33 +1,29 @@
-import { NextApiResponse, NextApiRequest } from "next";
 import * as prismic from "@prismicio/client";
+
+import { NextApiRequestLike, NextApiResponseLike } from "./types";
 
 /**
  * Configuration for `setPreviewData`.
  */
 export type SetPreviewDataConfig = {
 	/**
-	 * The `req` object from a Next.js API route. This is given as a parameter to
-	 * the API route.
+	 * The `req` object from a Next.js API route.
 	 *
-	 * @see Next.js API route docs: {@link https://nextjs.org/docs/api-routes/introduction}
+	 * @see Next.js API route docs: \<https://nextjs.org/docs/api-routes/introduction\>
 	 */
-	req: {
-		query: NextApiRequest["query"];
-		cookies: NextApiRequest["cookies"];
-	};
+	req: NextApiRequestLike;
 
 	/**
-	 * The `res` object from a Next.js API route. This is given as a parameter to
-	 * the API route.
+	 * The `res` object from a Next.js API route.
 	 *
-	 * @see Next.js API route docs: {@link https://nextjs.org/docs/api-routes/introduction}
+	 * @see Next.js API route docs: \<https://nextjs.org/docs/api-routes/introduction\>
 	 */
-	res: {
-		setPreviewData: NextApiResponse["setPreviewData"];
-	};
+	res: NextApiResponseLike;
 };
 
 /**
+ * **Only use this function in the Pages Directory (/pages).**
+ *
  * Set Prismic preview data for Next.js's Preview Mode.
  */
 export function setPreviewData({ req, res }: SetPreviewDataConfig): void {