+++ test/__setup__.ts
@@ -1,11 +1,14 @@
-import { beforeEach } from "vitest";
+import { beforeEach, vi } from "vitest";
 import { createMockFactory, MockFactory } from "@prismicio/mock";
+import { Headers } from "node-fetch";
 
 declare module "vitest" {
 	export interface TestContext {
 		mock: MockFactory;
 	}
 }
+
+vi.stubGlobal("Headers", Headers);
 
 beforeEach((ctx) => {
 	ctx.mock = createMockFactory({ seed: ctx.meta.name });